+++ skore/src/skore/_sklearn/evaluate.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING, cast
 
 from sklearn.base import BaseEstimator
 
+from skore import configuration
 from skore._sklearn._comparison.report import ComparisonReport
 from skore._sklearn._cross_validation.report import CrossValidationReport
 from skore._sklearn._estimator.report import EstimatorReport
@@ -145,7 +146,20 @@ def evaluate(
         )
 
     if isinstance(splitter, float):
-        splitter = TrainTestSplit(test_size=splitter)
+        # It's easier to make a 1-split CrossValidationReport
+        # and extract an EstimatorReport from it,
+        # than to make an EstimatorReport from scratch
+        with configuration(show_progress=False):
+            report = CrossValidationReport(
+                estimator,
+                X,
+                y,
+                data=data,
+                pos_label=pos_label,
+                splitter=TrainTestSplit(test_size=splitter),
+                n_jobs=n_jobs,
+            )
+        return report.estimator_reports_[0]
 
     report = CrossValidationReport(
         estimator,
@@ -156,6 +170,8 @@ def evaluate(
         splitter=splitter,
         n_jobs=n_jobs,
     )
+
+    # In case the splitter only produces one split, return the single estimator report
     if len(report.estimator_reports_) == 1:
         return report.estimator_reports_[0]
     return report